+++ CHANGELOG.md
@@ -27,6 +27,11 @@
 
   * [ExUnit.Formatter] Deprecate `format_time/2`, use `format_times/1` instead
 
+#### Mix
+
+  * [mix compile.leex] Require `:leex` to be added as a compiler to run the `leex` compiler
+  * [mix compile.yecc] Require `:yecc` to be added as a compiler to run the `yecc` compiler
+
 ## v1.15
 
 The CHANGELOG for v1.15 releases can be found [in the v1.15 branch](https://github.com/elixir-lang/elixir/blob/v1.15/CHANGELOG.md).